% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute Force Algorithm in R}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{Data.frame containing values and weights.}

\item{W}{Highest capacity}
}
\value{
Returns maximum knapsack value and respective elements
}
\description{
Knapsack problem solution using Brute Force Algorithm in R.
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
\references{
\href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
