% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy Algorithm in R}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Data.frame containing values and weights.}

\item{W}{Highest capacity}
}
\value{
Returns maximum knapsack value and respective elements
}
\description{
Knapsack problem solution using Greedy Algorithm in R.
}
\examples{
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)

}
\references{
\href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
