% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming in R}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Data.frame containing values and weights.}

\item{W}{Highest capacity}
}
\value{
Returns maximum knapsack value and respective elements
}
\description{
#' Knapsack problem solution using dynamic programming in R.
}
\examples{
knapsack_dynamic(x=knapsack_objects[1:1200,], W = 2000)

}
\references{
\href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
