% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack_objects}
\format{An object}
\source{
From the lab manual
}
\usage{
data(knapsack_objects)
}
\description{
knapsack_objects is a dataframe consisting of two variables values v and weights w.
}
\examples{
knapsack_dynamic(x=knapsack_objects[1:1200,], W = 2000)
}
\references{
\href{http://www.r-tutor.com/r-introduction/data-frame}{R tutorial}
}
\keyword{datasets}
