\name{RCourseLab6-package}
\docType{package}
\title{
\packageTitle{RCourseLab6}
}
\description{
\packageDescription{RCourseLab6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCourseLab6}
\packageIndices{RCourseLab6}
}
\author{
\packageAuthor{RCourseLab6}

Maintainer: \packageMaintainer{RCourseLab6}
}
\references{ \href{https://en.wikipedia.org/wiki/Knapsack_problem}{Dynamic Algorithm}
}
\keyword{ Greedy, Bruteforce, Dynamic, Knapsack }
\seealso{
\href{http://www.ida.liu.se/~732A94/info/courseinfo.en.shtml}{RCourseLab}
}
\examples{
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
greedy_knapsack(x = knapsack_objects[1:800,], W = 3500)
knapsack_dynamic(x=knapsack_objects[1:1200,], W = 2000)
}
