% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCourseLab6.R
\docType{package}
\name{RCourseLab6}
\alias{RCourseLab6}
\alias{RCourseLab6-package}
\title{A package for Lab 6 Assigments.}
\description{
Package contains of implementations of the algorithm.
Brute force search,Dynamic programming, Greedy heuristic. Also it contains an implementaion of 3 testsuits,
and vignette documentation.
}
\section{brute_force_knapsack}{

brute_force_knapsack() returns maximum knapsack value and respective elements accoring to brute force algorithm.
}

\section{knapsack_dynamic}{

knapsack_dynamic() returns maximum knapsack value and respective elements accoring to dynamic programming algorithm.
}

\section{greedy_knapsack}{

greedy_knapsack() returns maximum knapsack value and respective elements accoring to greedy algorithm.
}

\section{knapsack_objects}{

knapsack_objects contains the dataframe for the values and weights.
}

\seealso{
\link[RCourseLab6]{brute_force_knapsack}

\link[RCourseLab6]{knapsack_dynamic}

\link[RCourseLab6]{greedy_knapsack}

\link[RCourseLab6]{knapsack_objects}
}
